feat(rust): improve error and move it to ockam core
