+++ implementations/rust/ockam/ockam_core/Cargo.toml
@@ -1,9 +1,10 @@
 [package]
-name = "ockam_error"
+name = "ockam_core"
 version = "0.0.0"
 authors = ["Ockam Developers"]
 edition = "2018"
 license = "Apache-2.0"
 
-[lib]
-crate-type = ["rlib"]
+[features]
+default = ["std"]
+std = []
