+++ implementations/rust/ockam/ockam_core/src/error/no_std_error.rs
@@ -1,0 +1,14 @@
+#[derive(Debug)]
+pub struct Error {
+    code: u32,
+}
+
+impl Error {
+    pub fn new(code: u32) -> Self {
+        Self { code }
+    }
+
+    pub fn code(&self) -> u32 {
+        self.code
+    }
+}
