+++ implementations/rust/ockam/ockam_core/src/lib.rs
@@ -1,0 +1,7 @@
+// ---
+// #![no_std] if the standard library is not present.
+
+#![cfg_attr(not(feature = "std"), no_std)]
+
+mod error;
+pub use error::*;
